fix: close the API-golden gap and three ranking correctness bugs
The golden API file recorded declaration names only, so the assertion it
exists to make was still evadable: adding a method to engine.Scorer, a
parameter to Search, or retyping Document.Vector all left it passing. It
now records signatures, member types, interface methods and embedded
types, which covers every way a scorer can widen the shared contract.
Verified by mutation: an added Scorer.Stream line fails the test.

recency scored 2^(-age/HalfLife), which underflows float64 to exactly
zero past ~88 years. Every document older than that tied at zero and
TopK broke the tie on DocID, so insertion order silently replaced
recency — measured as [{0 0} {1 0} {2 0}] for documents aged 200, 100
and 90 years, returned in exactly the wrong order. Now 1/(1+age/HalfLife):
both forms decrease strictly with age, so the ordering is identical
wherever the exponential was representable, and the demo's fused output
is byte-identical. The score halves at one HalfLife and reaches 1/3, not
0.25, at two.

Index.Doc and Index.DocLen compared bounds with int(id). DocID is
uint32, so on a 32-bit build an id at or above 1<<31 wrapped negative,
passed the guard and panicked on the slice access instead of reporting
false as documented. Both now compare in uint64.

The BM25 posting scan checked ctx once per term and never inside the
loop, so a single-term query over a common term — the largest posting
list there is — finished the whole scan and sort and returned results
after the caller gave up. It now polls every 1024 postings and once more
before sorting.

Also corrects stale figures the previous round left behind: the line
counts in README and FINDINGS were two commits out of date, and two
comments still described the retracted "zero lines changed in engine/"
claim.

+++ pkg/engine/index.go
@@ -129,10 +129,15 @@ func (ix *Index) Len() int {
 // The returned Vector and Links alias index state and must not be modified,
 // the same contract as Lookup. Copying them here would allocate once per
 // document per query in the scorers that scan the whole corpus.
+//
+// The bound is compared in uint64, not int. DocID is uint32, so on a 32-bit
+// build int(id) wraps negative for an id at or above 1<<31: the guard would pass
+// and the slice access would panic, exactly where this is documented to return
+// false instead.
 func (ix *Index) Doc(id DocID) (Document, bool) {
 	ix.mu.RLock()
 	defer ix.mu.RUnlock()
-	if int(id) >= len(ix.docs) {
+	if uint64(id) >= uint64(len(ix.docs)) {
 		return Document{}, false
 	}
 	return ix.docs[id], true
@@ -156,11 +161,12 @@ func (ix *Index) Lookup(term string) []Posting {
 	return ix.postings[term]
 }
 
-// DocLen is the token count of a document, or 0 for an unknown id.
+// DocLen is the token count of a document, or 0 for an unknown id. The bound is
+// compared in uint64 for the reason given on Doc.
 func (ix *Index) DocLen(id DocID) int {
 	ix.mu.RLock()
 	defer ix.mu.RUnlock()
-	if int(id) >= len(ix.docLen) {
+	if uint64(id) >= uint64(len(ix.docLen)) {
 		return 0
 	}
 	return ix.docLen[id]
